Refactored and added documentation
