+++ src/components/landing-page/FeatEvents.js
@@ -6,7 +6,7 @@
 import React, { PropTypes } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import { fetchOneEvent } from '../../actions/';
+import { fetchOneEvent } from '../../redux/actions/';
 
 import Card from './EventCard';
